rafactor: 게임 상태 저장 및 복원 로직 개선
OmokGameState클래스 생성
